requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
et_xmlfile==2.0.0
html5lib==1.1
idna==3.10
jsonpath==0.82.2
lxml==5.3.1
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
py-mini-racer==0.6.0
PyMySQL==1.1.1
pypinyin==0.53.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
webencodings==0.5.1
xlrd==2.0.1